fix test merge conflict/move
